All elements have dark mode
