+++ src/components/App/App.component.jsx
@@ -1,5 +1,8 @@
-import React from 'react';
+import React, { useReducer } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
+
+import DarkModeReducer from '../../state/DarkModeReducer';
+import DarKModeContext from '../../state/DarkModeContext';
 
 import AuthProvider from '../../providers/Auth';
 import HomePage from '../../pages/Home';
@@ -10,6 +13,10 @@ import Private from '../Private';
 import Layout from '../Layout';
 
 function App() {
+  // const [darkModeState, darkModeDispatcher] = useReducer(DarkModeReducer, {
+  //   darkMode: false,
+  // })
+
   return (
     <BrowserRouter>
       <AuthProvider>
