+++ src/components/Layout/Layout.component.jsx
@@ -1,18 +1,11 @@
 import React, { useReducer, useEffect } from 'react';
-import styled from 'styled-components';
 import './Layout.sass';
-import { colors } from '../../utils/constants';
 
 import DarkModeReducer from '../../state/DarkModeReducer';
 import DarKModeContext from '../../state/DarkModeContext';
 import VideosContext from '../../state/VideosContext';
 import VideosReducer from '../../state/VideosReducer';
 import MenusComponent from '../Menus/Menus.component';
-
-const Container = styled.main`
-  background: ${colors.BG_SITE};
-  color: ${colors.FONT_SITE};
-`;
 
 function Layout({ children }) {
   const [darkModeState, darkModeDispatch] = useReducer(DarkModeReducer, {
@@ -35,14 +28,14 @@ function Layout({ children }) {
   const videosContextValue = { state: videosState, dispatch: videosDispatch };
 
   return (
-    <Container className="layout">
+    <main className={`layout ${darkModeState.darkMode && 'darkMode'}`}>
       <DarKModeContext.Provider value={darkModeContextValue}>
         <VideosContext.Provider value={videosContextValue}>
           <MenusComponent />
           {children}
         </VideosContext.Provider>
       </DarKModeContext.Provider>
-    </Container>
+    </main>
   );
 }
