+++ src/components/Layout/Layout.sass
@@ -4,3 +4,8 @@
   align-items: center
   margin: 0
   padding: 0
+  background-color: var(--bg-site)
+  overflow-y: scroll
+  &.darkMode
+    background-color: var(--dark-bg-site)
+    color: var(--dark-font-site)
