+++ src/components/Menus/TopMenu/DarkModeToggle/DarkModeToggle.sass
@@ -23,8 +23,8 @@
     right: 0
     bottom: 0
     background-color: rgba(0, 0, 0, 0.4)
-    -webkit-transition: .4s
-    transition: .4s
+    -webkit-transition: .5s
+    transition: .5s
     border-radius: 34px
 
   &-slider:before
@@ -34,15 +34,17 @@
     width: 20px
     left: -1px
     bottom: -2px
-    background-color: white
-    -webkit-transition: .4s
-    transition: .4s
+    background-color: #fff
+    -webkit-transition: .5s
+    transition: .5s
     border-radius: 50%
+    box-shadow: 0px 5px 18px -4px #000000
 
   &-input.checked + &-slider
-    background-color: rgba(255, 255, 255, 0.2)
+    background-color: rgba(25, 133, 123, 0.8)
 
   &-input.checked + &-slider:before
     -webkit-transform: translateX(26px)
     -ms-transform: translateX(26px)
     transform: translateX(14px)
+    background-color: rgba(25, 133, 123, 0.8)
