+++ src/components/Menus/TopMenu/Profile/Profile.sass
@@ -20,6 +20,13 @@
     border: 1px #4D4B4C
     margin: 0
     padding: 6px 0
+    background-color: var(--bg-site)
+    color: var(--font-site)
+    box-shadow: 2px 0px 10px 0px var(--box-shadow)
+    &.darkMode
+      background-color: var(--dark-bg-site)
+      color: var(--dark-font-site)
+      box-shadow: 1px 0px 4px 0px var(--dark-box-shadow)
     &-item
       text-align: left
       padding: 4px 14px
