+++ src/components/Menus/TopMenu/Search/Search.sass
@@ -13,6 +13,6 @@
     border: none
     font: inherit
     background: transparent
-    color: white
+    color: inherit
     &:focus
       outline: none
