+++ src/components/Menus/TopMenu/TopMenu.component.jsx
@@ -1,6 +1,5 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
-import { colors } from '../../../utils/constants';
 import './TopMenu.sass';
 
 import DarKModeContext from '../../../state/DarkModeContext';
@@ -8,10 +7,6 @@ import LeftMenuToggle from './LeftMenuToggle/LeftMenuToggle.component';
 import Search from './Search/Search.component';
 import DarkModeToggle from './DarkModeToggle/DarkModeToggle.component';
 import Profile from './Profile/Profile.component';
-
-const Container = styled.header`
-  background: ${colors.BG_TOP_MENU};
-`;
 
 const Space = styled.div`
   width: auto;
@@ -27,7 +22,7 @@ const TopMenu = (props) => {
     });
 
   return (
-    <Container className="TopMenu">
+    <div className={`TopMenu ${state.darkMode && 'darkMode'}`}>
       <div className="TopMenu-root">
         <LeftMenuToggle handleToggle={props.toggleLeftMenu} />
         <Search />
@@ -35,7 +30,7 @@ const TopMenu = (props) => {
         <DarkModeToggle darkMode={state.darkMode} handleToggle={handleDarkModeToggle} />
         <Profile />
       </div>
-    </Container>
+    </div>
   );
 };
