+++ src/components/Menus/TopMenu/TopMenu.sass
@@ -8,6 +8,10 @@
   flex-direction: column
   flex-shrink: 0
   padding: 12px 24px
+  background-color: var(--bg-top-menu)
+  transition: 0.5s
+  &.darkMode
+    background-color: var(--dark-bg-top-menu)
   &-root
     display: flex
     align-items: center
