+++ src/components/VideosList/VideosList.sass
@@ -20,6 +20,8 @@
     height: 345px
     margin: 10px
     box-shadow: 0px 2px 1px -1px rgba(0,0,0,0.2), 0px 1px 1px 0px rgba(0,0,0,0.14), 0px 1px 3px 0px rgba(0,0,0,0.12)
+    &.darkMode
+      box-shadow: 0px 2px 1px -1px rgba(255,255,255,0.2), 0px 1px 1px 0px rgba(255,255,255,0.14), 0px 1px 3px 0px rgba(255,255,255,0.12)
     &:hover
       opacity: 0.9
     &-thumbnail
@@ -38,11 +40,3 @@
         line-height: 1.6
         letter-spacing: 0.0075em
         margin-bottom: 0.35rem
-      p
-        margin: 0
-        font-size: 0.875rem
-        font-family: "Roboto", "Helvetica", "Arial", sans-serif
-        font-weight: 400
-        line-height: 1.43
-        letter-spacing: 0.01071em
-        color: rgba(0, 0, 0, 0.54)
